Merge changes from @firstyear
Rough initial clean up of the async code to wait_for_state

The design of async made a huge number of assupmtions about what the user wanted
and would include paths that would reboot/shutdown/create/destroy machines in
ways that weren't particularly obvious.

This removal is to keep the module simple. If the user wants to do something
dumb, we will fail fast rather than "assuming" and shooting people in the foot.

Additionally, by using wait_for_state, the user can choose what state they want
the machine to be in by the time the action is complete. For example, cloud init
can end with the machine up, or the machine being shutdown. Perhaps the user
wants to wait for the machine to be up? Perhaps they want it to be down before
they progess? This will allow that to happen.

Added polling and wait for state. Need to rewrite calling logic

Final fixes to add_tags usage